addes likes, comments, comment likes

+++ controllers/feed.js
@@ -6,20 +6,47 @@ const { validationResult } = require('express-validator/check');
 const io = require('../socket');
 const Post = require('../models/post');
 const User = require('../models/user');
+const Like = require('../models/like');
 
 exports.getPosts = async (req, res, next) => {
   const currentPage = req.query.page || 1;
-  const perPage = 2;
+  const perPage = 10;
   try {
     const totalItems = await Post.find().countDocuments();
     const posts = await Post.find()
       .populate('creator')
+      .populate({
+        path: 'likes',
+        populate: { 
+          path: 'userId',
+          select: ['name', '_id']
+        }
+      })
+      .populate({
+        path: 'comments',
+        populate: [{ 
+          path: 'userId',
+          select: ['name', '_id'],
+        },{
+          path: 'likes',
+          populate: { 
+            path: 'userId',
+            select: ['name', '_id'],
+          }
+        }]
+      })
+      // .populate({
+      //   path: 'comments',
+      //   populate: { 
+      //     path: 'likes',
+      //     select: ['name', '_id'],
+      //   }
+      // })
       .sort({ createdAt: -1 })
       .skip((currentPage - 1) * perPage)
       .limit(perPage);
 
     res.status(200).json({
-      message: 'Fetched posts successfully.',
       posts: posts,
       totalItems: totalItems
     });
@@ -32,27 +59,28 @@ exports.getPosts = async (req, res, next) => {
 };
 
 exports.createPost = async (req, res, next) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    const error = new Error('Validation failed, entered data is incorrect.');
-    error.statusCode = 422;
-    throw error;
-  }
-  if (!req.file) {
-    const error = new Error('No image provided.');
-    error.statusCode = 422;
-    throw error;
-  }
-  const imageUrl = req.file.path;
-  const title = req.body.title;
-  const content = req.body.content;
-  const post = new Post({
-    title: title,
-    content: content,
-    imageUrl: imageUrl,
-    creator: req.userId
-  });
   try {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      const error = new Error('Validation failed, entered data is incorrect.');
+      error.statusCode = 422;
+      throw error;
+    }
+    if (!req.file) {
+      const error = new Error('No image provided.');
+      error.statusCode = 422;
+      throw error;
+    }
+
+    const imageUrl = req.file.path;
+    const title = req.body.title;
+    const content = req.body.content;
+    const post = new Post({
+      title: title,
+      content: content,
+      imageUrl: imageUrl,
+      creator: req.userId
+    });
     await post.save();
     const user = await User.findById(req.userId);
     user.posts.push(post);
@@ -63,7 +91,6 @@ exports.createPost = async (req, res, next) => {
       creator: { _id: user._id, name: user.name }
     });
     res.status(201).json({
-      message: 'Post created successfully!',
       post: post,
       creator: { _id: user._id, name: user.name }
     });
@@ -77,14 +104,22 @@ exports.createPost = async (req, res, next) => {
 
 exports.getPost = async (req, res, next) => {
   const postId = req.params.postId;
-  const post = await Post.findById(postId);
+  const post = await Post.findById(postId)
+    .populate('creator')
+    .populate({
+      path: 'likes',
+      populate: { 
+        path: 'userId',
+        select: ['name', '_id']
+      }
+    });
   try {
     if (!post) {
       const error = new Error('Could not find post.');
       error.statusCode = 404;
       throw error;
     }
-    res.status(200).json({ message: 'Post fetched.', post: post });
+    res.status(200).json({ post: post });
   } catch (err) {
     if (!err.statusCode) {
       err.statusCode = 500;
@@ -96,23 +131,25 @@ exports.getPost = async (req, res, next) => {
 exports.updatePost = async (req, res, next) => {
   const postId = req.params.postId;
   const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    const error = new Error('Validation failed, entered data is incorrect.');
-    error.statusCode = 422;
-    throw error;
-  }
-  const title = req.body.title;
-  const content = req.body.content;
-  let imageUrl = req.body.image;
-  if (req.file) {
-    imageUrl = req.file.path;
-  }
-  if (!imageUrl) {
-    const error = new Error('No file picked.');
-    error.statusCode = 422;
-    throw error;
-  }
   try {
+    if (!errors.isEmpty()) {
+      const error = new Error('Validation failed, entered data is incorrect.');
+      error.statusCode = 422;
+      throw error;
+    }
+    const title = req.body.title;
+    const content = req.body.content;
+    let imageUrl = req.body.image;
+
+    if (req.file) {
+      imageUrl = req.file.path;
+    }
+    if (!imageUrl) {
+      const error = new Error('No file picked.');
+      error.statusCode = 422;
+      throw error;
+    }
+
     const post = await Post.findById(postId).populate('creator');
     if (!post) {
       const error = new Error('Could not find post.');
@@ -134,9 +171,8 @@ exports.updatePost = async (req, res, next) => {
     io.getIO().emit('posts', {
       action: 'update',
       post: { ...result._doc },
-      creator: { _id: user._id, name: user.name }
     });
-    res.status(200).json({ message: 'Post updated!', post: result });
+    res.status(200).json(result);
   } catch (err) {
     if (!err.statusCode) {
       err.statusCode = 500;
@@ -160,7 +196,12 @@ exports.deletePost = async (req, res, next) => {
       error.statusCode = 403;
       throw error;
     }
-    // Check logged in user
+    if (post.creator._id.toString() !== req.userId) {
+      const error = new Error('Not authorized!');
+      error.statusCode = 403;
+      throw error;
+    }
+
     clearImage(post.imageUrl);
     await Post.findByIdAndRemove(postId);
     const user = await User.findById(req.userId);
@@ -168,9 +209,9 @@ exports.deletePost = async (req, res, next) => {
     await user.save();
     io.getIO().emit('posts', {
       action: 'delete',
-      post:postId,
+      post: postId,
     });
-    res.status(200).json({ message: 'Deleted post.' });
+    res.status(200).json(postId);
   } catch (err) {
     if (!err.statusCode) {
       err.statusCode = 500;
@@ -178,6 +219,46 @@ exports.deletePost = async (req, res, next) => {
     next(err);
   }
 };
+
+exports.postLike = async (req, res, next) => {
+  const postId = req.params.postId;
+  try {
+    const post = await Post.findById(postId);
+    if (!post) {
+      const error = new Error('Post was not exist');
+      error.statusCode = 422;
+      throw error;
+    }
+    const likeExist = await Like.findOne({
+      postId: postId,
+      userId: req.userId
+    });
+    if(likeExist) {
+      await Like.findByIdAndRemove(likeExist._id);
+      post.likes.pull(likeExist._id);
+      await post.save();
+      res.status(200).json('unliked');
+      return next();
+    }
+    const likeData = new Like({
+      postId: post._id,
+      userId: req.userId
+    });
+    const like = await likeData.save();
+    post.likes.push(like._id);
+    await post.save();
+    const likePopulated = await Like.findById(like._id).populate({
+      path: 'userId',
+      select: ['name', '_id']
+    });
+    res.status(200).json(likePopulated);
+  } catch (err) {
+    if (!err.statusCode) {
+      err.statusCode = 500;
+    }
+    next(err);
+  }
+}
 
 const clearImage = filePath => {
   filePath = path.join(__dirname, '..', filePath);
